+++ packages/analytics/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/Availity/availity-react/compare/@availity/analytics@1.1.2...@availity/analytics@1.1.3) (2019-11-20)
+
+**Note:** Version bump only for package @availity/analytics
+
+
+
+
+
+## [1.1.2](https://github.com/Availity/availity-react/compare/@availity/analytics@1.1.1...@availity/analytics@1.1.2) (2019-11-01)
+
+**Note:** Version bump only for package @availity/analytics
+
+
+
+
+
 ## [1.1.1](https://github.com/Availity/availity-react/compare/@availity/analytics@1.1.0...@availity/analytics@1.1.1) (2019-10-24)
 
 **Note:** Version bump only for package @availity/analytics